+++ terraform/main.tf
@@ -6,7 +6,7 @@ terraform {
       version = "~> 5.32"
     }
   }
-  required_version = "~> 1.10.4"
+  required_version = ">= 1.10, < 2.0.0"
 }
 
 data "aws_region" "current" {}